backend/requirements.txt
django == 5.1.7
djangorestframework == 3.15.2
pytest-django == 4.10.0